pyproject.toml
[project]
name = "next"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.2",
    "fastapi>=0.115.0",
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.0",
    "pysbd>=0.3.4",
    "torch>=2.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "uvicorn>=0.30.0",
]
